# packages/contracts/openapi/kitchenflow-v1.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "description": "Stable machine-readable KitchenFlow error code.",
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Field-level validation messages keyed by request field name.",
            "type": "object"
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "description": "Correlation identifier suitable for support without exposing private request data.",
            "type": "string"
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "errorCode",
          "traceId"
        ],
        "type": "object"
      },
...
      "RecipeGenerationSessionResponse": {
        "description": "HTTP projection of one cook-now generation session.",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RecipeCandidateResponse"
            },
            "type": [
              "null",
              "array"
            ]
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "sessionId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "status",
          "candidates",
          "failureReason"
        ],
        "type": "object"
      },
      "RecipeSummaryResponse": {
        "description": "HTTP projection of one owned recipe summary.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "mealTypes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "recipeId": {
            "format": "uuid",
            "type": "string"
          },
          "servings": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "recipeId",
          "name",
          "mealTypes",
          "servings",
          "createdAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "kitchenflowSession": {
        "description": "Backend-managed, secure HttpOnly session cookie. Browser code must not read or store OIDC tokens.",
        "in": "cookie",
        "name": "__Host-kitchenflow-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "license": {
      "name": "PolyForm Noncommercial License 1.0.0",
      "url": "https://polyformproject.org/licenses/noncommercial/1.0.0/"
    },
    "title": "KitchenFlow.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/recipes": {
      "get": {
        "operationId": "getApiV1Recipes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeSummaryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/recipes",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
    "/api/v1/recipes/generation-sessions": {
      "post": {
        "operationId": "postApiV1RecipesGeneration-sessions",
        "parameters": [
          {
            "description": "Required CSRF token issued by GET /api/v1/session.",
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required UUID key for cook-now generation or selection replay.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeGenerationSessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "rateLimitFailure": {
                    "summary": "The bounded mutation rate was exceeded.",
                    "value": {
                      "detail": "The request rate limit was exceeded.",
                      "errorCode": "rate_limit_exceeded",
                      "status": 429,
                      "title": "Too Many Requests",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The mutation rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "POST /api/v1/recipes/generation-sessions",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
    "/api/v1/recipes/generation-sessions/{sessionId}": {
      "get": {
        "operationId": "getApiV1RecipesGeneration-sessionsSessionId",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeGenerationSessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/recipes/generation-sessions/{sessionId}",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
    "/api/v1/recipes/generation-sessions/{sessionId}/selection": {
      "post": {
        "operationId": "postApiV1RecipesGeneration-sessionsSessionIdSelection",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Required CSRF token issued by GET /api/v1/session.",
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required UUID key for cook-now generation or selection replay.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectCandidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetailResponse"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetailResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unsupportedMediaType": {
                    "summary": "The request content type is unsupported.",
                    "value": {
                      "detail": "The request content type is not supported.",
                      "errorCode": "unsupported_media_type",
                      "status": 415,
                      "title": "Unsupported Media Type",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request content type is not supported."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "rateLimitFailure": {
                    "summary": "The bounded mutation rate was exceeded.",
                    "value": {
                      "detail": "The request rate limit was exceeded.",
                      "errorCode": "rate_limit_exceeded",
                      "status": 429,
                      "title": "Too Many Requests",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The mutation rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "POST /api/v1/recipes/generation-sessions/{sessionId}/selection",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
    "/api/v1/recipes/{recipeId}": {
      "get": {
        "operationId": "getApiV1RecipesRecipeId",
        "parameters": [
          {
            "in": "path",
            "name": "recipeId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetailResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "authenticationFailure": {
                    "summary": "The backend-managed session is absent or invalid.",
                    "value": {
                      "detail": "Authentication is required.",
                      "errorCode": "authentication_required",
                      "status": 401,
                      "title": "Unauthorized",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "unexpectedFailure": {
                    "summary": "An unexpected error was safely redacted.",
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "errorCode": "unexpected_error",
                      "status": 500,
                      "title": "Internal Server Error",
                      "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An unexpected server error occurred."
          }
        },
        "security": [
          {
            "kitchenflowSession": []
          }
        ],
        "summary": "GET /api/v1/recipes/{recipeId}",
        "tags": [
          "KitchenFlow.Api"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Current KitchenFlow backend origin. Local development defaults to https://localhost:7443.",
      "url": "/"
    }
  ],
...
}